--- Person.rb
@@ -1,25 +1,0 @@
-class Person
-  def initialize(name = 'Unknown', age = 0, parent_permission : true)
-    @id = Random.rand(1..1000)
-    @name = name
-    @age = age
-    @parent_permission = parent_permission
-  end
-
-  private
-
-  def of_age?
-    !(age >= 18)
-  end
-
-  public
-
-  def can_use_services?
-    return unless of_age? && @parent_permission
-
-    true
-  end
-
-  attr_accessor :name, :age
-  attr_reader :id
-end
