--- Student.rb
@@ -1,11 +1,0 @@
-require_relative 'person'
-
-class Student < Person
-  def initialize(name = 'Unknown', age = 0, parent_permission: true, classroom)
-    super(name, age, parent_permission: parent_permission)
-    @classroom = classroom
-  end
-
-  def play_hooky
-    "¯\\_(ツ)_/¯"
-  end
